requirements.txt
requests
pytest
voluptuous
pytest-voluptuous
allure-pytest
curlify
selene
selenium